+++ client/components/cards/resultCard.styl
@@ -19,3 +19,6 @@
 
 .result-card-context-list
   margin-bottom: 0.7rem
+
+.result-card-block-wrapper
+  display: inline-block
