+++ i18n/en.i18n.json
@@ -876,6 +876,7 @@
   "label-not-found": "Label '%s' not found.",
   "label-color-not-found": "Label color %s not found.",
   "user-username-not-found": "Username '%s' not found.",
+  "comment-not-found": "Card with comment containing text '%s' not found.",
   "globalSearch-title": "Search All Boards",
   "no-cards-found": "No Cards Found",
   "one-card-found": "One Card Found",
@@ -901,6 +902,9 @@
   "operator-modified": "modified",
   "operator-sort": "sort",
   "operator-comment": "comment",
+  "predicate-archived": "archived",
+  "predicate-active": "active",
+  "predicate-overdue": "overdue",
   "operator-unknown-error": "%s is not an operator",
   "operator-number-expected": "operator __operator__ expected a number, got '__value__'",
   "operator-sort-invalid": "sort of '%s' is invalid",
@@ -911,12 +915,16 @@
   "globalSearch-instructions-operator-board": "`__operator_board__:title` - cards in boards matching the specified title",
   "globalSearch-instructions-operator-list": "`__operator_list__:title` - cards in lists matching the specified title",
   "globalSearch-instructions-operator-swimlane": "`__operator_swimlane__:title` - cards in swimlanes matching the specified title",
+  "globalSearch-instructions-operator-comment": "`__operator_comment__:text` - cards with with a comment containing *text*.",
   "globalSearch-instructions-operator-label": "`__operator_label__:color` `__operator_label__:name` - cards that have a label matching the given color or name",
   "globalSearch-instructions-operator-hash": "`__operator_label_abbrev__label` - shorthand for `__operator_label__:label`",
   "globalSearch-instructions-operator-user": "`__operator_user__:username` - cards where the specified user is a *member* or *assignee*",
   "globalSearch-instructions-operator-at": "`__operator_user_abbrev__username` - shorthand for `user:username`",
   "globalSearch-instructions-operator-member": "`__operator_member__:username` - cards where the specified user is a *member*",
   "globalSearch-instructions-operator-assignee": "`__operator_assignee__:username` - cards where the specified user is an *assignee*",
+  "globalSearch-instructions-operator-due": "`__operator_due__:n` - cards which are due *n* days from now.  `__operator_due__:__predicate_overdue__ lists all ",
+  "globalSearch-instructions-operator-created": "`__operator_created__:n` - cards which which were created *n* days ago",
+  "globalSearch-instructions-operator-modified": "`__operator_modified__:n` - cards which which were modified *n* days ago",
   "globalSearch-instructions-notes-1": "Multiple operators may be specified.",
   "globalSearch-instructions-notes-2": "Similar operators are *OR*ed together.  Cards that match any of the conditions will be returned.\n`__operator_list__:Available __operator_list__:Blocked` would return cards contained in any list named *Blocked* or *Available*.",
   "globalSearch-instructions-notes-3": "Differing operators are *AND*ed together.  Only cards that match all of the differing operators are returned.\n`__operator_list__:Available __operator_label__:red` returns only cards in the list *Available* with a *red* label.",
