+++ models/cardComments.js
@@ -125,7 +125,10 @@ CardComments.textSearch = (userId, textArray) => {
 
   const comments = CardComments.find(selector);
   // eslint-disable-next-line no-console
-  console.log('count:', comments.count());
+  // console.log('count:', comments.count());
+  // eslint-disable-next-line no-console
+  // console.log('cards with comments:', comments.map(com => { return com.cardId }));
+
   return comments;
 };
